+++ kazoo/testing/common.py
@@ -169,15 +169,13 @@ log4j.appender.ROLLINGFILE.File=""" + to_java_compatible_path(
             jars.extend(glob(os.path.join(
                 self.install_path,
                 "lib/*.jar")))
+            # support for different file locations on Debian/Ubuntu
             jars.extend(glob(os.path.join(
                 self.install_path,
                 "log4j-*.jar")))
             jars.extend(glob(os.path.join(
                 self.install_path,
-                "slf4j-api-*.jar")))
-            jars.extend(glob(os.path.join(
-                self.install_path,
-                "slf4j-log4j*.jar")))
+                "slf4j-*.jar")))
         else:
             # Development build (plain `ant`)
             jars = glob((os.path.join(